requirements.txt
fastmcp==2.11
google-genai>=0.2.0
numpy>=1.24.0
requests>=2.28.0